(#88)feat: 기업 정보 없음 UI 구현했습니다.
