+++ src/asset/sass/pages/searchPage/companyRegistPage.scss
@@ -1,0 +1,9 @@
+.input-field {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 10px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+}
