+++ src/asset/sass/pages/searchPage/searchPage.scss
@@ -1,0 +1,41 @@
+span.result-data-failed {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-top: 10%;
+  padding: 60px;
+}
+
+.failed-text {
+  letter-spacing: -1px;
+  margin-top: 10%;
+  font-size: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-weight: 500;
+}
+
+.failed-text2 {
+  color: #7e7e7e;
+  letter-spacing: -1px;
+  margin-top: 2%;
+  margin-bottom: 10%;
+  font-size: 13px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.company-registration {
+  letter-spacing: -1px;
+  margin-top: -1%;
+  color: #ff8d1d;
+  cursor: pointer;
+}
+
+.registration-container {
+  display: flex;
+  flex-direction: row;
+}
