+++ src/router/allRouter.js
@@ -2,14 +2,26 @@ import React, { useEffect } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import { ACCESS_TOKEN } from '../components/pages/loginPage/constants/index.js';
 
+/* 메인 페이지 관련 impot */
 import MainPage from '../components/pages/mainPage/mainPage';
+
+/* 사용자 등록 페이지 관련  */
 import LoginPage from '../components/pages/loginPage/loginPage';
 import NicknamePage from '../components/pages/loginPage/nicknamePage';
-import MyPage from '../components/pages/myPage/myPage';
-import NotificationPage from '../components/pages/notificationPage/notificationPage';
-import TokenManagement from '../components/global/token/tokenManagement';
+
+/* 사용자 참여 페이지 관련 */
 import Store from '../components/pages/storePage/storePage';
-import SearchPage from '../components/pages/mainPage/searchPage.jsx';
+import MyPage from '../components/pages/myPage/myPage';
+import CompanyRegistPage from '../components/pages/searchPage/companyRegistPage.jsx';
+import NotificationPage from '../components/pages/notificationPage/notificationPage';
+
+/* 질문 페이지 관련 */
+import SearchPage from '../components/pages/searchPage/searchPage.jsx';
+
+/* 그 외 유틸  */
+import TokenManagement from '../components/global/token/tokenManagement';
+
+/* ===================================================================  */
 
 const AllRouter = () => {
   useEffect(() => {
@@ -25,12 +37,16 @@ const AllRouter = () => {
     <div className="app">
       <Routes>
         <Route path="/" element={<MainPage />} />
+        <Route path="/auth/token" element={<TokenManagement />} />
+
         <Route path="/login" element={<LoginPage />} />
         <Route path="/login/member-info" element={<NicknamePage />} />
+
+        <Route path="/store" element={<Store />} />
         <Route path="/mypage" element={<MyPage />} />
         <Route path="/notification" element={<NotificationPage />} />
-        <Route path="/auth/token" element={<TokenManagement />} />
-        <Route path="/store" element={<Store />} />
+        <Route path="/company-regist" element={<CompanyRegistPage />} />
+
         <Route path="/search-company" element={<SearchPage />} />
       </Routes>
     </div>
